requirements.txt
Flask
python-dotenv
python-pptx
requests
SQLAlchemy>=2.0
psycopg2-binary
